+++ packages/user-profile-widget/rollup.config.app.mjs
@@ -24,7 +24,6 @@ export default {
           process.env.DESCOPE_PROJECT_ID || '',
         ),
         DESCOPE_BASE_URL: JSON.stringify(process.env.DESCOPE_BASE_URL || ''),
-        DESCOPE_TENANT: JSON.stringify(process.env.DESCOPE_TENANT || ''),
         DESCOPE_WIDGET_ID: JSON.stringify(process.env.DESCOPE_WIDGET_ID || ''),
       },
     }),