+++ packages/libs/sdk-component-drivers/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.2.12](https://github.com/descope/descope-js/compare/sdk-component-drivers-0.2.11...sdk-component-drivers-0.2.12) (2024-05-11)
+
+### Dependency Updates
+
+* `sdk-helpers` updated to version `0.1.31`
+## [0.2.11](https://github.com/descope/descope-js/compare/sdk-component-drivers-0.2.10...sdk-component-drivers-0.2.11) (2024-05-07)
+
+### Dependency Updates
+
+* `sdk-helpers` updated to version `0.1.30`
 ## [0.2.10](https://github.com/descope/descope-js/compare/sdk-component-drivers-0.2.9...sdk-component-drivers-0.2.10) (2024-05-02)
 
 ### Dependency Updates