+++ packages/user-management-widget/CHANGELOG.md
@@ -2,6 +2,27 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.4.57](https://github.com/descope/descope-js/compare/user-management-widget-0.4.56...user-management-widget-0.4.57) (2024-05-12)
+
+### Dependency Updates
+
+* `web-js-sdk` updated to version `1.10.47`
+## [0.4.56](https://github.com/descope/descope-js/compare/user-management-widget-0.4.55...user-management-widget-0.4.56) (2024-05-11)
+
+### Dependency Updates
+
+* `web-js-sdk` updated to version `1.10.46`
+* `sdk-helpers` updated to version `0.1.31`
+* `sdk-mixins` updated to version `0.2.13`
+* `sdk-component-drivers` updated to version `0.2.12`
+## [0.4.55](https://github.com/descope/descope-js/compare/user-management-widget-0.4.54...user-management-widget-0.4.55) (2024-05-07)
+
+### Dependency Updates
+
+* `web-js-sdk` updated to version `1.10.45`
+* `sdk-helpers` updated to version `0.1.30`
+* `sdk-mixins` updated to version `0.2.12`
+* `sdk-component-drivers` updated to version `0.2.11`
 ## [0.4.54](https://github.com/descope/descope-js/compare/user-management-widget-0.4.53...user-management-widget-0.4.54) (2024-05-02)
 
 ### Dependency Updates